[fail] The property defined on the Config interface made the target contain an invalid filter.
This was a bit puzzling since Mark had a working system and mine failed. It turned out that Mark had
no parameter names compiled into the class. Therefore, the name of the 'tls'
parameter became `$000` in the DS XML. Since this
did not match the tls.target name in the Config
interface the default value was not set. However,
I had the proper name and therefore the tls.target
became the empty string and made the code fail.

Removed this settings. Need to discuss how to
work with the communication provider that this bundle is.

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>